+++ kernel/mem/bitmap.c
@@ -14,7 +14,7 @@
 #include "printk.h"
 
 /* Initialize the memory bitmap */
-void bitmap_init(Bitmap *bitmap, uint8_t *buffer, unsigned long size)
+void bitmap_init(Bitmap *bitmap, uint8_t *buffer, size_t size)
 {
 	bitmap->buffer = buffer;
 	bitmap->length = size * 8;
@@ -22,57 +22,57 @@ void bitmap_init(Bitmap *bitmap, uint8_t *buffer, unsigned long size)
 }
 
 /* Get memory bitmap */
-int bitmap_get(const Bitmap *bitmap, unsigned long index)
+int bitmap_get(const Bitmap *bitmap, size_t index)
 {
-	unsigned long word_index = index / 8;
-	unsigned long bit_index = index % 8;
+	size_t word_index = index / 8;
+	size_t bit_index = index % 8;
 	return (bitmap->buffer[word_index] >> bit_index) & 1;
 }
 
 /* Setting the memory bitmap */
-void bitmap_set(Bitmap *bitmap, unsigned long index, int value)
+void bitmap_set(Bitmap *bitmap, size_t index, int value)
 {
-	unsigned long word_index = index / 8;
-	unsigned long bit_index = index % 8;
+	size_t word_index = index / 8;
+	size_t bit_index = index % 8;
 	if (value) {
-		bitmap->buffer[word_index] |= ((unsigned long)1 << bit_index);
+		bitmap->buffer[word_index] |= ((size_t)1 << bit_index);
 	} else {
-		bitmap->buffer[word_index] &= ~((unsigned long)1 << bit_index);
+		bitmap->buffer[word_index] &= ~((size_t)1 << bit_index);
 	}
 }
 
 /* Set the memory bitmap range */
-void bitmap_set_range(Bitmap *bitmap, unsigned long start, unsigned long end, int value)
+void bitmap_set_range(Bitmap *bitmap, size_t start, size_t end, int value)
 {
-	unsigned long start_word = (start + 7) / 8;
-	unsigned long end_word = end / 8;
+	size_t start_word = (start + 7) / 8;
+	size_t end_word = end / 8;
 	if (start >= end || start >= bitmap->length) {
 		return;
 	}
-	for (unsigned long i = start; i < start_word * 8 && i < end; i++) {
+	for (size_t i = start; i < start_word * 8 && i < end; i++) {
 		bitmap_set(bitmap, i, value);
 	}
 	if (start_word > end_word) {
 		return;
 	}
 	if (start_word <= end_word) {
-		unsigned long fill_value = value ? (unsigned long)-1 : 0;
-		for (unsigned long i = start_word; i < end_word; i++) {
+		size_t fill_value = value ? (size_t)-1 : 0;
+		for (size_t i = start_word; i < end_word; i++) {
 			bitmap->buffer[i] = fill_value;
 		}
 	}
-	for (unsigned long i = end_word * 8; i < end; i++) {
+	for (size_t i = end_word * 8; i < end; i++) {
 		bitmap_set(bitmap, i, value);
 	}
 }
 
 /* Memory bitmap search range */
-unsigned long bitmap_find_range(const Bitmap *bitmap, unsigned long length, int value)
+size_t bitmap_find_range(const Bitmap *bitmap, size_t length, int value)
 {
-	unsigned long count = 0, start_index = 0;
+	size_t count = 0, start_index = 0;
 	uint8_t byte_match = value ? (uint8_t)-1 : 0;
-	for (unsigned long byte_idx = 0; byte_idx < bitmap->length / 8; byte_idx++) {
-		unsigned long byte = bitmap->buffer[byte_idx];
+	for (size_t byte_idx = 0; byte_idx < bitmap->length / 8; byte_idx++) {
+		size_t byte = bitmap->buffer[byte_idx];
 		if (byte == !byte_match) {
 			count = 0;
 		} else if (byte == byte_match) {
@@ -87,7 +87,7 @@ unsigned long bitmap_find_range(const Bitmap *bitmap, unsigned long length, int 
 				return start_index;
 			}
 		} else {
-			for (unsigned long bit = 0; bit < 8; bit++) {
+			for (size_t bit = 0; bit < 8; bit++) {
 				int bit_value = (byte >> bit) & 1;
 				if (bit_value == value) {
 					if (count == 0) {
@@ -103,5 +103,5 @@ unsigned long bitmap_find_range(const Bitmap *bitmap, unsigned long length, int 
 			}
 		}
 	}
-	return (unsigned long)-1;
+	return (size_t)-1;
 }
